Pservice Module

+++ application/modules/pservice/controllers/Pservice.php
@@ -8,14 +8,30 @@ class Pservice extends MX_Controller {
     function __construct() {
         parent::__construct();
         $this->load->model('pservice_model');
-        if (!$this->ion_auth->in_group(array('admin', 'Nurse', 'Doctor', 'Accountant'))) {
-       
+        $group_permission = $this->ion_auth->get_users_groups()->row();
+
+        if ($group_permission->name == 'admin' || $group_permission->name == 'Patient' || $group_permission->name == 'Doctor' || $group_permission->name == 'Nurse' || $group_permission->name == 'Pharmacist' || $group_permission->name == 'Laboratorist' || $group_permission->name == 'Accountant' || $group_permission->name == 'Receptionist' || $group_permission->name == 'members') {
+
+            $this->pers = array();
+            $this->permission_access_group_explode = array();
+        } else {
+            $this->pers = explode(',', $group_permission->description);
+
+            $this->db->where('group_id', $group_permission->id);
+            $query = $this->db->get('permission_access_group')->row();
+            $permission_access_group = $query->permission_access;
+            $this->permission_access_group_explode = explode('***', $permission_access_group);
+        }
+        if ($this->ion_auth->in_group(array('Laboratorist', 'Receptionist', 'Patient', 'pharmacist'))) {
+
             redirect('home/permission');
         }
     }
 
     public function index() {
- if (!$this->ion_auth->in_group(array('admin'))) {
+
+        if (!$this->ion_auth->in_group(array('admin')) && !in_array('Bed', $this->pers)) {
+
             redirect('home/permission');
         }
         $data['pservices'] = $this->pservice_model->getPservice();
@@ -26,7 +42,9 @@ class Pservice extends MX_Controller {
     }
 
     public function addNewView() {
-         if (!$this->ion_auth->in_group(array('admin'))) {
+
+        if (!$this->ion_auth->in_group(array('admin')) && !in_array('Bed', $this->pers)) {
+
             redirect('home/permission');
         }
         $data['pservices'] = $this->pservice_model->getPservice();
@@ -37,18 +55,18 @@ class Pservice extends MX_Controller {
     }
 
     public function addNew() {
-        
-       $active= $this->input->post('active');
-    if(empty($active)){
-        $active='0';
-    }else{
-        $active='1';
-    }
-      $id = $this->input->post('id');
+
+        $active = $this->input->post('active');
+        if (empty($active)) {
+            $active = '0';
+        } else {
+            $active = '1';
+        }
+        $id = $this->input->post('id');
         $name = $this->input->post('name');
         $code = $this->input->post('code');
         $alpha_code = $this->input->post('alpha_code');
-        
+
         $price = $this->input->post('price');
         $referential_price = $this->input->post('referential_price');
         $special_price = $this->input->post('special_price');
@@ -105,7 +123,9 @@ class Pservice extends MX_Controller {
     }
 
     function editPservice() {
-         if (!$this->ion_auth->in_group(array('admin'))) {
+
+        if (!$this->ion_auth->in_group(array('admin')) && !in_array('Bed', $this->pers)) {
+
             redirect('home/permission');
         }
         $data = array();
@@ -117,7 +137,9 @@ class Pservice extends MX_Controller {
     }
 
     function editPserviceByJason() {
-         if (!$this->ion_auth->in_group(array('admin'))) {
+
+        if (!$this->ion_auth->in_group(array('admin')) && !in_array('Bed', $this->pers)) {
+
             redirect('home/permission');
         }
         $id = $this->input->get('id');
@@ -126,7 +148,9 @@ class Pservice extends MX_Controller {
     }
 
     function delete() {
-         if (!$this->ion_auth->in_group(array('admin'))) {
+
+        if (!$this->ion_auth->in_group(array('admin')) && !in_array('Bed', $this->pers)) {
+
             redirect('home/permission');
         }
         $data = array();
@@ -137,9 +161,8 @@ class Pservice extends MX_Controller {
     }
 
     function getPserviceList() {
-         if (!$this->ion_auth->in_group(array('admin'))) {
-            redirect('home/permission');
-        }
+
+
         $requestData = $_REQUEST;
         $start = $requestData['start'];
         $limit = $requestData['length'];
@@ -161,12 +184,40 @@ class Pservice extends MX_Controller {
 
         //  $data['patients'] = $this->patient_model->getVisitor();
         $i = 0;
+        $permis = '';
+        $permis_1 = '';
+        $permis_2 = '';
+        foreach ($this->permission_access_group_explode as $perm) {
+            $perm_explode = array();
+            // $permis = '';
+            // $permis_1 = '';
+            //$permis_2 = '';
+            $perm_explode = explode(",", $perm);
+            if (in_array('2', $perm_explode) && $perm_explode[0] == 'Bed') {
+                $permis = 'ok';
+                //  break;
+            }
+            if (in_array('1', $perm_explode) && $perm_explode[0] == 'Bed') {
+                $permis_1 = 'ok';
+                //  break;
+            }
+            if (in_array('3', $perm_explode) && $perm_explode[0] == 'Bed') {
+                $permis_2 = 'ok';
+                //  break;
+            }
+        }
         foreach ($data['pservices'] as $pservice) {
             $i = $i + 1;
+            $option2 = '';
+            $option1 = '';
+            if ($this->ion_auth->in_group(array('admin')) || $permis == 'ok') {
 
-            $option1 = '<button type="button" class="btn btn-info btn-xs btn_width editbutton" data-toggle="modal" data-id="' . $pservice->id . '"><i class="fa fa-edit"> ' . lang('edit') . '</i></button>';
+                $option1 = '<button type="button" class="btn btn-info btn-xs btn_width editbutton" data-toggle="modal" data-id="' . $pservice->id . '"><i class="fa fa-edit"> ' . lang('edit') . '</i></button>';
+            }
 
-            $option2 = '<a class="btn btn-info btn-xs btn_width delete_button" href="pservice/delete?id=' . $pservice->id . '" onclick="return confirm(\'Are you sure you want to delete this item?\');"><i class="fa fa-trash"> </i></a>';
+            if ($this->ion_auth->in_group(array('admin')) || $permis_2 == 'ok') {
+                $option2 = '<a class="btn btn-info btn-xs btn_width delete_button" href="pservice/delete?id=' . $pservice->id . '" onclick="return confirm(\'Are you sure you want to delete this item?\');"><i class="fa fa-trash"> </i></a>';
+            }
             if (!empty($pservice->price)) {
                 $price = $pservice->price;
             } else {
